use a trait to share deployment details logic
